backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
fastapi==0.115.0
uvicorn[standard]==0.32.0
beautifulsoup4==4.14.3
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
deprecation==2.1.0
exceptiongroup==1.3.1
fsspec==2025.10.0
gotrue==2.9.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.26.0
hyperframe==6.1.0
idna==3.11
markdown-it-py==3.0.0
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.0
numpy==2.0.2
packaging==25.0
pandas==2.3.3
postgrest==0.16.11
propcache==0.4.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.3.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-telegram-bot==20.8
pytz==2025.2
realtime==1.0.6
requests==2.32.5
rich==14.2.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
storage3==0.7.7
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.6.0
supabase-auth==2.27.2
supabase-functions==2.27.2
supafunc==0.5.1
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==12.0
yarl==1.22.0
